refactor(runtime): reuse blocking pool completion helper
Isomorphism: repeated blocking-pool metamorphic test blocks now share sorted_completed_task_ids. Submission order, wait semantics, completion collection, and sort order are unchanged for the affected relations.

Fresh-eyes fixes: mr_spawn_blocking_cancellation_states now installs a runtime-backed Cx before asserting pool-backed spawn_blocking cancellation semantics, waits on real blocking-thread completion for the soft-cancel case, and cancellation_visualizer removes a no-op u32 conversion found by test-aware clippy.

Verification: rustfmt --edition 2024 --check src/runtime/blocking_pool/metamorphic.rs src/observability/cancellation_visualizer.rs; git diff --check; rch exec -- cargo test --target-dir /tmp/cargo-target-asupersync-blocking-pool-pass010-test3 -p asupersync --lib mr_spawn_blocking_cancellation_states; rch exec -- cargo test --target-dir /tmp/cargo-target-asupersync-blocking-pool-pass010-test3 -p asupersync --lib runtime::blocking_pool::metamorphic; rch exec -- cargo check --target-dir /tmp/cargo-target-asupersync-pass010-check -p asupersync --lib; rch exec -- cargo clippy --target-dir /tmp/cargo-target-asupersync-pass010-clippy-tests -p asupersync --lib --tests -- -D warnings.

+++ src/observability/cancellation_visualizer.rs
@@ -809,7 +809,7 @@ mod tests {
             cancel_reason: "User(test)".to_string(),
             cancel_kind: "User".to_string(),
             timestamp: std::time::SystemTime::UNIX_EPOCH + elapsed_since_prev,
-            elapsed_since_start: elapsed_since_prev * (u32::from(step_id) + 1),
+            elapsed_since_start: elapsed_since_prev * (step_id + 1),
             elapsed_since_prev,
             depth,
             parent_entity: parent_entity.map(str::to_string),
